rename folder -> folders

+++ src/infra/src/table/entity/folders.rs
@@ -3,10 +3,10 @@
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "folder")]
+#[sea_orm(table_name = "folders")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     #[sea_orm(unique)]
     pub folder_id: String,
     pub org: String,
